add naive reuse detector to memdb
A common problem with go-memdb is:

1. Fetch *Foo from go-memdb
2. Change Foo.Bar from 2 => 3
3. Insert Foo into go-memdb

Step 2 races with readers of the same object from go-memdb as go-memdb
shares pointers with all readers. The proper approach is:

1. Fetch
2. Copy
3. Update copy
4. Insert copy

This panics on the 1st case.

However I'm not sure `cannot insert the same pointer twice` is a safe
invariant to assume as the existing test suite does not work with this
enabled.

Try it yourself with:

go test -tags memdbrace